RadioLocal/requirements.txt
fastapi==0.110.2
fastapi-health
uvicorn==0.29.0
pydantic==2.7.1
pydantic_core==2.18.2
pydantic-settings==2.3.1
pymongo==4.6.1
cryptography==41.0.7
PyJWT==2.8.0
python-dotenv==1.0.1